Fix: 댓글 평점 정수형 변경

+++ src/main/java/com/greedy/mokkoji/api/comment/dto/response/CommentResponse.java
@@ -5,12 +5,12 @@ import java.time.LocalDateTime;
 public record CommentResponse(
         Long id,
         String content,
-        Double rate,
+        Integer rate,
         boolean isModified,
         LocalDateTime time,
         boolean isWriter
 ) {
-    public static CommentResponse of(final Long id, final String content, final Double rate, final boolean isModified, final LocalDateTime time, final boolean isWriter) {
+    public static CommentResponse of(final Long id, final String content, final Integer rate, final boolean isModified, final LocalDateTime time, final boolean isWriter) {
         return new CommentResponse(id, content, rate, isModified, time, isWriter);
     }
 }
